# agent-builder-platform/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Builder Platform API",
...
    "version": "1.0.0",
    "contact": {
      "name": "Agent Builder Platform Team",
      "email": "support@agent-builder-platform.com",
      "url": "https://agent-builder-platform.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "servers": [
      {
        "url": "http://localhost:8000",
        "description": "Local development server"
      },
      {
        "url": "https://api.agent-builder-platform.com",
        "description": "Production server"
      },
      {
        "url": "https://staging-api.agent-builder-platform.com",
        "description": "Staging server"
      }
    ]
  },
  "paths": {
...
    "/api/sessions": {
      "post": {
        "tags": [
          "Sessions"
        ],
        "summary": "Create Session",
        "description": "Create a new agent creation session\n\n- **user_id**: Optional user identifier (anonymous if not provided)\n- **experience_level**: User's experience level (beginner, intermediate, advanced, expert)\n\nReturns session details including session_id for subsequent API calls",
        "operationId": "create_session_api_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}": {
      "get": {
        "tags": [
          "Sessions"
        ],
        "summary": "Get Session",
        "description": "Retrieve session details by session ID\n\n- **session_id**: Unique session identifier\n\nReturns complete session state including current phase and progress",
        "operationId": "get_session_api_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                },
                "examples": {
                  "session_created": {
                    "summary": "Session created successfully",
                    "value": {
                      "session_id": "session-abc123def456",
                      "status": "active",
                      "current_phase": "requirements",
                      "created_at": "2025-10-06T10:00:00Z",
                      "updated_at": "2025-10-06T10:00:00Z",
                      "experience_level": "intermediate"
                    }
                  },
                  "agent_created": {
                    "summary": "Agent workflow created",
                    "value": {
                      "agent_id": "agent-xyz789abc123",
                      "session_id": "session-abc123def456",
                      "status": "requirements",
                      "progress_percentage": 0.0,
                      "created_at": "2025-10-06T10:05:00Z",
                      "use_case": "customer_support_chatbot",
                      "description": "Build an AI-powered chatbot...",
                      "experience_level": "intermediate",
                      "vector_search_enabled": true
                    }
                  },
                  "requirements_analysis": {
                    "summary": "Requirements analysis complete",
                    "value": {
                      "session_id": "session-abc123def456",
                      "use_case_analysis": {
                        "primary_use_case": "chatbot",
                        "complexity": "medium",
                        "estimated_duration": "35_minutes"
                      },
                      "service_recommendations": [
                        {
                          "service_name": "AWS Lambda",
                          "reasoning": "Serverless compute for chatbot logic with automatic scaling",
                          "estimated_cost": "$5-10/month",
                          "confidence": 0.97
                        },
                        {
                          "service_name": "Amazon Lex",
                          "reasoning": "Pre-built conversational AI with NLU capabilities",
                          "estimated_cost": "$15-20/month",
                          "confidence": 0.95
                        }
                      ],
                      "cost_analysis": {
                        "estimated_monthly_cost": "$25-35",
                        "breakdown": {
                          "compute": "$5-10",
                          "ai_services": "$15-20",
                          "storage": "$2-3",
                          "data_transfer": "$3-5"
                        },
                        "optimization_opportunities": [
                          "Use Lambda free tier (1M requests/month)",
                          "Enable S3 Intelligent-Tiering for logs"
                        ]
                      },
                      "security_analysis": {
                        "recommendations": [
                          "Enable encryption at rest for DynamoDB",
                          "Use IAM roles instead of access keys",
                          "Implement API Gateway throttling"
                        ],
                        "compliance_considerations": [
                          "GDPR",
                          "SOC2"
                        ]
                      },
                      "clarifying_questions": [
                        "What is the expected number of concurrent users?",
                        "Do you need multi-language support?",
                        "Should the chatbot integrate with existing systems?"
                      ],
                      "confidence_score": 0.97,
                      "next_steps": [
                        "Answer clarifying questions",
                        "Review service recommendations",
                        "Proceed to architecture design"
                      ]
                    }
                  },
                  "test_results": {
                    "summary": "Test execution results",
                    "value": {
                      "agent_id": "agent-xyz789abc123",
                      "test_execution_id": "test-exec-123456",
                      "timestamp": "2025-10-06T10:30:00Z",
                      "overall_status": "passed",
                      "tests_executed": {
                        "security": 8,
                        "performance": 5,
                        "cost": 3,
                        "integration": 6,
                        "load": 4
                      },
                      "summary": {
                        "total_tests": 26,
                        "passed": 24,
                        "failed": 1,
                        "warnings": 1
                      },
                      "production_readiness_score": 0.92,
                      "confidence_score": 0.97,
                      "recommendations": [
                        "Enable CloudWatch alarms for Lambda errors",
                        "Add DynamoDB auto-scaling policies",
                        "Implement request validation middleware"
                      ]
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Sessions"
        ],
        "summary": "Delete Session",
        "description": "Delete a session and cleanup associated data\n\n- **session_id**: Unique session identifier\n\nReturns confirmation of deletion",
        "operationId": "delete_session_api_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSessionResponse"
                },
                "examples": {
                  "session_created": {
                    "summary": "Session created successfully",
                    "value": {
                      "session_id": "session-abc123def456",
                      "status": "active",
                      "current_phase": "requirements",
                      "created_at": "2025-10-06T10:00:00Z",
                      "updated_at": "2025-10-06T10:00:00Z",
                      "experience_level": "intermediate"
                    }
                  },
                  "agent_created": {
                    "summary": "Agent workflow created",
                    "value": {
                      "agent_id": "agent-xyz789abc123",
                      "session_id": "session-abc123def456",
                      "status": "requirements",
                      "progress_percentage": 0.0,
                      "created_at": "2025-10-06T10:05:00Z",
                      "use_case": "customer_support_chatbot",
                      "description": "Build an AI-powered chatbot...",
                      "experience_level": "intermediate",
                      "vector_search_enabled": true
                    }
                  },
                  "requirements_analysis": {
                    "summary": "Requirements analysis complete",
                    "value": {
                      "session_id": "session-abc123def456",
                      "use_case_analysis": {
                        "primary_use_case": "chatbot",
                        "complexity": "medium",
                        "estimated_duration": "35_minutes"
                      },
                      "service_recommendations": [
                        {
                          "service_name": "AWS Lambda",
                          "reasoning": "Serverless compute for chatbot logic with automatic scaling",
                          "estimated_cost": "$5-10/month",
                          "confidence": 0.97
                        },
                        {
                          "service_name": "Amazon Lex",
                          "reasoning": "Pre-built conversational AI with NLU capabilities",
                          "estimated_cost": "$15-20/month",
                          "confidence": 0.95
                        }
                      ],
                      "cost_analysis": {
                        "estimated_monthly_cost": "$25-35",
                        "breakdown": {
                          "compute": "$5-10",
                          "ai_services": "$15-20",
                          "storage": "$2-3",
                          "data_transfer": "$3-5"
                        },
                        "optimization_opportunities": [
                          "Use Lambda free tier (1M requests/month)",
                          "Enable S3 Intelligent-Tiering for logs"
                        ]
                      },
                      "security_analysis": {
                        "recommendations": [
                          "Enable encryption at rest for DynamoDB",
                          "Use IAM roles instead of access keys",
                          "Implement API Gateway throttling"
                        ],
                        "compliance_considerations": [
                          "GDPR",
                          "SOC2"
                        ]
                      },
                      "clarifying_questions": [
                        "What is the expected number of concurrent users?",
                        "Do you need multi-language support?",
                        "Should the chatbot integrate with existing systems?"
                      ],
                      "confidence_score": 0.97,
                      "next_steps": [
                        "Answer clarifying questions",
                        "Review service recommendations",
                        "Proceed to architecture design"
                      ]
                    }
                  },
                  "test_results": {
                    "summary": "Test execution results",
                    "value": {
                      "agent_id": "agent-xyz789abc123",
                      "test_execution_id": "test-exec-123456",
                      "timestamp": "2025-10-06T10:30:00Z",
                      "overall_status": "passed",
                      "tests_executed": {
                        "security": 8,
                        "performance": 5,
                        "cost": 3,
                        "integration": 6,
                        "load": 4
                      },
                      "summary": {
                        "total_tests": 26,
                        "passed": 24,
                        "failed": 1,
                        "warnings": 1
                      },
                      "production_readiness_score": 0.92,
                      "confidence_score": 0.97,
                      "recommendations": [
                        "Enable CloudWatch alarms for Lambda errors",
                        "Add DynamoDB auto-scaling policies",
                        "Implement request validation middleware"
                      ]
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSessionRequest": {
        "properties": {
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "User identifier"
          },
          "experience_level": {
            "$ref": "#/components/schemas/ExperienceLevel",
            "description": "User experience level",
            "default": "beginner"
          }
        },
        "type": "object",
        "title": "CreateSessionRequest"
      },
      "DeleteSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "deleted_at": {
            "type": "string",
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message",
          "deleted_at"
        ],
        "title": "DeleteSessionResponse"
      },
...
      "ExperienceLevel": {
        "type": "string",
        "enum": [
          "beginner",
          "intermediate",
          "advanced",
          "expert"
        ],
        "title": "ExperienceLevel"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus"
          },
          "current_phase": {
            "$ref": "#/components/schemas/WorkflowPhase"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "experience_level": {
            "$ref": "#/components/schemas/ExperienceLevel"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "current_phase",
          "created_at",
          "updated_at",
          "experience_level"
        ],
        "title": "SessionResponse"
      },
      "SessionStatus": {
        "type": "string",
        "enum": [
          "active",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "SessionStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowPhase": {
        "type": "string",
        "enum": [
          "requirements",
          "architecture",
          "implementation",
          "testing",
          "deployment"
        ],
        "title": "WorkflowPhase"
      }
    }
  }
}